+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to Rust's notion of
   - `FieldExt::from_u64` (use `From<u64> for ff::PrimeField` instead).
   - `FieldExt::{from_bytes, read, to_bytes, write}`
     (use `ff::PrimeField::{from_repr, to_repr}` instead).
+  - `FieldExt::rand` (use `ff::Field::random` instead).
 
 ## [0.2.1] - 2021-09-17
 ### Changed